+++ CHANGELOG.md
@@ -1,0 +1,27 @@
+Changelog for EGCG-Project-Management
+=====================================
+
+0.2.3
+-----
+ - Refactor of automatic review and data deletion
+ - Update data deletion to use Archive management from EGCG-Core
+ 
+
+0.2.2  
+-----
+
+0.2.1b
+------
+
+0.2.1a
+------
+
+0.2.1
+-----
+
+0.2
+---
+This version adds a script to generate the project report.
+
+0.1
+---
