+++ data_deletion/delivered_data.py
@@ -3,170 +3,10 @@ import uuid
 from datetime import datetime
 from os import listdir
 from os.path import join, isdir
-from egcg_core import rest_communication, util, clarity
-from egcg_core.constants import ELEMENT_SAMPLE_INTERNAL_ID, ELEMENT_PROJECT_ID, ELEMENT_RUN_NAME, ELEMENT_LANE, \
-    ELEMENT_SAMPLE_EXTERNAL_ID
+from egcg_core import rest_communication, util
 from egcg_core.config import cfg
-from egcg_core.app_logging import AppLogger
-from data_deletion import Deleter
-from data_deletion.archive_management import is_archived, ArchivingError, release_file_from_lustre
-
-
-def get_file_list_size(file_list):
-    """
-    Get the total size of all files. Collapses them by inodes to avoid counting hard links more than once.
-    Also descends into directories recursively.
-    """
-    def files_by_inode(_file_list):
-        inode2file = {}
-        for f in _file_list:
-            if os.path.isdir(f):
-                inode2file.update(files_by_inode(os.path.join(f, s) for s in os.listdir(f)))
-            else:
-                inode2file[os.stat(f).st_ino] = f
-        return inode2file
-
-    inode2file = files_by_inode(file_list)
-    return sum(os.stat(f).st_size for f in inode2file.values())
-
-
-class ProcessedSample(AppLogger):
-    def __init__(self, sample_data):
-        self.raw_data_dir = cfg['data_deletion']['fastqs']
-        self.processed_data_dir = cfg['data_deletion']['processed_data']
-        self.delivered_data_dir = cfg['data_deletion']['delivered_data']
-        self.sample_data = sample_data
-        self._release_date = None
-        self._release_folders = None
-        self._run_elements = None
-        self._files_to_purge = None
-        self._files_to_remove_from_lustre = None
-        self._size_of_files = None
-
-    @property
-    def release_date(self):
-        if not self._release_date:
-            self._release_date = clarity.get_sample_release_date(self.sample_id)
-        return self._release_date
-
-    @property
-    def sample_id(self):
-        return self.sample_data[ELEMENT_SAMPLE_INTERNAL_ID]
-
-    @property
-    def project_id(self):
-        return self.sample_data[ELEMENT_PROJECT_ID]
-
-    @property
-    def external_sample_id(self):
-        return self.sample_data[ELEMENT_SAMPLE_EXTERNAL_ID]
-
-    def _find_fastqs_for_run_element(self, run_element):
-        return util.find_fastqs(
-            join(self.raw_data_dir, run_element[ELEMENT_RUN_NAME]),
-            run_element[ELEMENT_PROJECT_ID],
-            run_element[ELEMENT_SAMPLE_INTERNAL_ID],
-            lane=run_element[ELEMENT_LANE]
-        )
-
-    @property
-    def run_elements(self):
-        if self._run_elements is None:
-            self._run_elements = rest_communication.get_documents(
-                'run_elements', quiet=True, where={ELEMENT_SAMPLE_INTERNAL_ID: self.sample_id}, all_pages=True
-            )
-        return self._run_elements
-
-    def _raw_data_files(self):
-        all_fastqs = []
-        for e in self.run_elements:
-            assert e[ELEMENT_SAMPLE_INTERNAL_ID] == self.sample_id
-            fastqs = self._find_fastqs_for_run_element(e)
-            if fastqs:
-                all_fastqs.extend(fastqs)
-            else:
-                self.warning(
-                    'No fastqs found for run %s lane %s sample %s',
-                    e[ELEMENT_RUN_NAME],
-                    e[ELEMENT_LANE],
-                    self.sample_id
-                )
-        return all_fastqs
-
-    def _processed_data_files(self):
-        files_to_delete = []
-        # TODO: first check what type of analysis was perfomed on this data to know exactly which files need to be deleted
-        files_to_search = [
-            '{ext_s_id}_R1.fastq.gz', '{ext_s_id}_R2.fastq.gz',
-            '{ext_s_id}.bam', '{ext_s_id}.bam.bai',
-            '{ext_s_id}.vcf.gz', '{ext_s_id}.vcf.gz.tbi',
-            '{ext_s_id}.g.vcf.gz', '{ext_s_id}.g.vcf.gz.tbi'
-        ]
-        for f in files_to_search:
-            file_to_delete = util.find_file(
-                self.processed_data_dir,
-                self.project_id, self.sample_id,
-                f.format(ext_s_id=self.external_sample_id)
-            )
-            if file_to_delete:
-                files_to_delete.append(file_to_delete)
-        return files_to_delete
-
-    @property
-    def released_data_folder(self):
-        if self._release_folders is None:
-            self._release_folders = util.find_files(self.delivered_data_dir, self.project_id, '*', self.sample_id)
-        if len(self._release_folders) != 1:
-            self.warning(
-                'Found %s deletable directories for sample %s: %s',
-                len(self._release_folders),
-                self.sample_id,
-                self._release_folders
-            )
-        else:
-            return self._release_folders[0]
-
-    @property
-    def files_to_purge(self):
-        if not self._files_to_purge:
-            self._files_to_purge = []
-            release_folder = self.released_data_folder
-            if release_folder:
-                self._files_to_purge.append(release_folder)
-        return self._files_to_purge
-
-    @property
-    def files_to_remove_from_lustre(self):
-        if not self._files_to_remove_from_lustre:
-            self._files_to_remove_from_lustre = []
-            raw_files = self._raw_data_files()
-            if raw_files:
-                self._files_to_remove_from_lustre.extend(raw_files)
-
-            processed_files = self._processed_data_files()
-            if processed_files:
-                self._files_to_remove_from_lustre.extend(processed_files)
-            for f in self._files_to_remove_from_lustre:
-                if not is_archived(f):
-                    raise ArchivingError('File %s is not archived so cannot be released from Lustre' % f)
-        return self._files_to_remove_from_lustre
-
-    @property
-    def size_of_files(self):
-        if not self._size_of_files:
-            self._size_of_files = get_file_list_size(self.files_to_purge)
-            self._size_of_files += get_file_list_size(self.files_to_remove_from_lustre)
-        return self._size_of_files
-
-    def mark_as_deleted(self):
-        rest_communication.patch_entry(
-            'samples',
-            {'data_deleted': 'on lustre'},
-            'sample_id', self.sample_id
-        )
-
-    def __repr__(self):
-        return self.sample_id + ' (%s)' % self.release_date
+from data_deletion import Deleter, ProcessedSample
+from egcg_core.archive_management import release_file_from_lustre
 
 
 class DeliveredDataDeleter(Deleter):
