+++ requirements.txt
@@ -1,3 +1,4 @@
-git+https://github.com/EdinburghGenomics/EGCG-Core.git@v0.6.6
+EGCG-Core>=0.6.7
 jinja2==2.8
 weasyprint==0.30
+cached-property>=1.3.0
