+++ requirements_no_pdf.txt
@@ -1,2 +1,3 @@
-git+https://github.com/EdinburghGenomics/EGCG-Core.git@v0.5
+EGCG-Core>=0.6.7
 jinja2==2.8
+cached-property>=1.3.0
